Added Role Admin and control user routes. Seeder users and roles

+++ app/Http/Controllers/UserController.php
@@ -2,26 +2,31 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\UserRequest;
 use App\Http\Resources\UserResource;
+use App\Http\Requests\UserRequest;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use App\Models\User;
+use Auth;
 
 class UserController extends Controller
 {
+
     public function index(Request $request)
     {
         $query = User::query();
 
+//        Filter name
         if($name = $request->get('name', '')){
             $query->where('name', 'like', "%$name%");
         }
 
+//        Filter email
         if($email = $request->get('email', '')){
             $query->where('email', 'like', "%$email%");
         }
 
+//        orderBy
         if($orderBy = $request->get('sortBy')){
             $sortDesct = $request->get('sortDesc', 'false');
             if($sortDesct === 'true'){
@@ -30,9 +35,10 @@ class UserController extends Controller
                 $query->orderBy($orderBy);
             }
         } else{
-            $query->orderBy('id', 'desc');
+            $query->orderBy('id', 'asc');
         }
 
+//        pagination
         if($request->get('page', '')){
             $perPage = $request->get('perPage', 10);
             $users = $query->paginate($perPage);
@@ -72,11 +78,49 @@ class UserController extends Controller
 
     public function destroy(User $user): JsonResponse
     {
-        $user->delete();
+//        Can't delete himself
+        if($user->id !== Auth::user()->id){
+            $user->delete();
 
-        return response()->json([
-            'message' => 'Successfully deleted user!'
-        ]);
+            return response()->json([
+                'message' => 'Successfully deleted user!'
+            ]);
+        } else{
+            return response()->json([
+                'message' => "You can't do this action!"
+            ], 403);
+        }
+    }
+
+    /**
+     * This role can assign and revoke the 'Admin' role
+     */
+    public function changeAdminRole(User $user): JsonResponse
+    {
+//        Can't change role to himself
+        if($user->id !== Auth::user()->id){
+
+            if($user->hasRole('Admin')){
+
+                $user->removeRole('Admin');
+                return response()->json([
+                    'message' => 'Successfully revoked Admin role to user!'
+                ]);
+
+            } else{
+
+                $user->assignRole('Admin');
+                return response()->json([
+                    'message' => 'Successfully assigned Admin role to user!'
+                ]);
+
+            }
+
+        } else{
+            return response()->json([
+                'message' => "You can't do this action!"
+            ], 403);
+        }
     }
 
 }
